Sentiment/requirements.txt
newsapi-python
pymongo
python-dotenv
protobuf
regex